Addition of a CLI option to only run tests on specific backend/device combinations, for fast prototyping. (#1708)
Changing the CLI test configuration to add flags for device/backend selection.

Without these flags, PyTest will check all available backends and for each of the all devices, which is very thorough but also takes a lot of time while being unnecessary for many debugging purposes (with bugs that can be addressed also e.g. on NumPy-only, which is much quicker to test).