Globally consistent signatures for the `element` methods of ODL spaces.
At one point a `device` argument was considered, but that does not make mu sense
given that devices are fixed in the space and all elements of a space must have
the same device.

Furthermore, the `cast` argument was intended to select whether only lists of
subspace-elements were allowed as input to the `element` method of product
spaces (see 5bf1347c9). This was quite ad-hoc, specific to product space
for no very convincing reason, and of dubious use.

Meanwhile, the `copy` argument is sensible to disallow potential performance
degradation. Unlike with `cast`, it is not so straightforward to add type-rigidity
with an additional assertion for this, since a copy can be necessary for
numerous different reasons.

These changes were based on parts of https://github.com/odlgroup/odl/pull/1710.